refactor app file and create display module
